fix: 펀딩 목록 페이지 수정
